third_party/blink/web_tests/fast/dom/NodeIterator/detach-no-op-expected.txt

This test checks that NodeIterator.detach() is a no-op.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS iterator.nextNode().id is "a"
PASS iterator.detach() did not throw exception.
PASS iterator.nextNode().id is "b"
PASS successfullyParsed is true

TEST COMPLETE
